DO NOT MERGE: Clean up analog joystick decoding thingies
THIS IS ONLY A DRAFT. DO NOT MERGE YET.

+++ src/client/joystick_controller.h
@@ -149,7 +149,7 @@ public:
 		return m_axes_vals[axis];
 	}
 
-	s16 getAxisWithoutDead(JoystickAxis axis);
+	float getAxisWithoutDead(JoystickAxis axis);
 
 	f32 doubling_dtime;
